requirement.txt
absl-py==0.13.0
anyio==3.3.1
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.1.3
conda==4.10.3
conda-package-handling==1.7.3
cryptography==3.4.7
cycler==0.10.0
dataclasses==0.6
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
easydict==1.9
entrypoints==0.3
filelock==3.9.0
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.40.0
h5py==2.10.0
huggingface-hub==0.12.0
idna==2.10
imageio==2.9.0
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyterlab==3.1.12
jupyterlab-language-pack-zh-CN==0.0.1.dev0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.8.8
notebook==6.4.4
numpy==1.23.5
oauthlib==3.1.1
opencv-python==4.1.2.30
packaging==21.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pip==21.1.3
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
psutil==5.9.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycosat==0.6.3
pycparser==2.20
Pygments==2.10.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-crf==0.7.2
pytz==2021.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
regex==2022.10.31
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
ruamel-yaml-conda==0.15.100
sacremoses==0.0.53
scikit-image==0.16.2
scikit-learn==1.1.3
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
seqeval==1.2.2
setuptools==52.0.0.post20210125
six==1.16.0
sniffio==1.2.0
supervisor==4.2.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
tokenizers==0.10.3
torch==1.7.0+cu110
TorchCRF==1.0.8
torchvision==0.8.1+cu110
tornado==6.1
tqdm==4.43.0
traitlets==5.1.0
transformers==4.11.3
typing-extensions==3.10.0.2
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
wheel==0.36.2
widgetsnbextension==3.5.1
